+++ app/src/views/Bambuser/index.ts
@@ -1,1 +1,1 @@
-export * from './Bambuser'
+export * from './BambuserView'
