+++ app/src/views/Bambuser/utils.ts
@@ -1,6 +1,5 @@
 import gql from 'graphql-tag'
 
-import { productQuery } from '../../../pages/products/[productSlug]'
 import { Sentry } from '../../../src/services/sentry'
 import {
   request,
@@ -26,7 +25,6 @@ interface Response {
 type mesgProps = (url: string) => void
 
 export const queryByHandle = async (handle: string) => {
-  // productQuery
   const query = gql`
     query ProductsPageQuery($handle: String) {
       allShopifyProduct(
@@ -96,130 +94,4 @@ export const onMessage = (callback: mesgProps) => {
 
 export const insideIframe = (): boolean => {
   return window !== window.parent
-}
-
-export const findVariant = (product: ShopifyProduct, vid: string) => {
-  return product?.variants?.find((variant) => {
-    return variant?.title === vid
-  })
-}
-
-export const hydrateSize = (
-  product: ShopifyProduct,
-  s: any,
-  color: string | null = '',
-) => {
-  const sizes = product?.options?.find((option) => {
-    return option?.name === 'Size'
-  })
-  if (sizes && sizes.values && sizes.values.length > 0) {
-    return sizes?.values?.map((size) => {
-      let key
-      if (color) {
-        key = `${color}${DELIMITER}${size?.value}`
-      } else {
-        key = `${size?.value}`
-      }
-      const variant = findVariant(product, key)
-      if (
-        variant &&
-        variant.sourceData &&
-        variant.sourceData.availableForSale
-      ) {
-        const price = variant.sourceData.priceV2
-        return s()
-          .name(size?.value)
-          .price((pr) =>
-            pr.currency(price?.currencyCode).current(price?.amount),
-          )
-          .sku(variant.shopifyVariantID)
-      }
-    })
-  } else {
-    const key = `${color}`
-    const variant = findVariant(product, key)
-    if (variant && variant.sourceData && variant.sourceData.availableForSale) {
-      const price = variant.sourceData.priceV2
-      return [
-        s()
-          .name(product.title)
-          .price((pr) =>
-            pr.currency(price?.currencyCode).current(price?.amount),
-          )
-          .sku(variant.shopifyVariantID),
-      ]
-    }
-  }
-}
-
-export const hydrate = (product: ShopifyProduct, v: any) => {
-  const colors = product?.options?.find((option) => {
-    return option?.name === 'Color'
-  })
-
-  const sizes = product?.options?.find((option) => {
-    return option?.name === 'Size'
-  })
-
-  const arraymove = (arr, fromIndex, toIndex) => {
-    var element = arr[fromIndex]
-    arr.splice(fromIndex, 1)
-    arr.splice(toIndex, 0, element)
-  }
-
-  let collection = colors?.values
-  // if ( collection && collection.length > 0) {
-  //   arraymove(collection, 2, 0)
-  // }
-
-  if (collection && collection.length > 0) {
-    return collection.map((color) => {
-      let variant,
-        key,
-        image: string[] = []
-
-      const sizeValue =
-        sizes && sizes.values && sizes.values.length > 0
-          ? sizes.values[0]?.value
-          : ''
-      if (colors && sizeValue) {
-        key = `${color?.value}${DELIMITER}${sizeValue}`
-      } else {
-        key = `${color?.value}${sizeValue}`
-      }
-      variant = findVariant(product, key)
-
-      if (variant && variant.sourceData) {
-        image.push(variant.sourceData?.image?.w800)
-      }
-
-      return v()
-        .attributes((a) => a.colorName(color?.value))
-        .imageUrls(image)
-        .sku(variant.shopifyVariantID)
-        .name('color?.value')
-        .sizes((s) => hydrateSize(product, s, color?.value))
-    })
-  } else {
-    let variant,
-      key,
-      image: string[] = []
-    if (sizes) {
-      key =
-        sizes && sizes.values && sizes.values.length > 0
-          ? sizes.values[0]?.value
-          : ''
-      variant = findVariant(product, key)
-      if (variant && variant.sourceData) {
-        image.push(variant.sourceData?.image?.w800)
-      }
-    }
-    return [
-      v()
-        .sku(variant.shopifyVariantID)
-        .name('No Color Option')
-        .imageUrls(image)
-        .sizes((s) => hydrateSize(product, s)),
-    ]
-  }
 }
